Fix bug with finding versioned toolchains
